+++ src/SimulationRunner.cc
@@ -21,9 +21,11 @@
 
 #include "ignition/gazebo/components/Model.hh"
 #include "ignition/gazebo/components/Name.hh"
+#include "ignition/gazebo/components/Sensor.hh"
 #include "ignition/gazebo/components/World.hh"
 #include "ignition/gazebo/Events.hh"
 #include "ignition/gazebo/SdfEntityCreator.hh"
+#include "ignition/gazebo/Util.hh"
 #include "network/NetworkManagerPrimary.hh"
 
 using namespace ignition;
@@ -568,6 +570,22 @@ void SimulationRunner::LoadPlugins(const Entity _entity,
     {
       entity = this->entityCompMgr.EntityByComponents(
           components::Name(plugin.EntityName()), components::World());
+    }
+    else if ("sensor" == plugin.EntityType())
+    {
+      // TODO(louise) Use scoped names for models and worlds too
+      auto sensors = this->entityCompMgr.EntitiesByComponents(
+          components::Sensor());
+
+      for (auto sensor : sensors)
+      {
+        if (scopedName(sensor, this->entityCompMgr, "::", false) ==
+            plugin.EntityName())
+        {
+          entity = sensor;
+          break;
+        }
+      }
     }
     else
     {